+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Thunderdome Changelog
 
 ## Unreleased Changes
+* Added `Arena::with_capacity` for preallocating space.
 
 ## 0.1.0 (2020-09-02)
 * Initial release